+++ R/ggplot_scatter_fun.R
@@ -163,8 +163,8 @@ ggplot_scatter <- function(data,
                            y_title = "[Y title]",
                            caption = "",
                            font_family = "Helvetica",
-                           font_size_title = 11,
-                           font_size_body = 10,
+                           font_size_title = NULL,
+                           font_size_body = NULL,
                            wrap_title = 70,
                            wrap_subtitle = 80,
                            wrap_x_title = 50,
@@ -182,6 +182,15 @@ ggplot_scatter <- function(data,
   
   if (!is.numeric(x_var_vector)) stop("Please use a numeric x variable for a scatterplot")
   if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a scatterplot")
+  
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
   
   if (is.null(pal)) pal <- pal_snz
   
@@ -383,8 +392,8 @@ ggplot_scatter_col <-
            caption = "",
            legend_labels = NULL,
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -405,6 +414,15 @@ ggplot_scatter_col <-
     
     if (!is.numeric(x_var_vector)) stop("Please use a numeric x variable for a scatterplot")
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a scatterplot")
+    
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
+    }
     
     if (is.null(col_method)) {
       if (!is.numeric(col_var_vector)) col_method <- "category"
@@ -640,8 +658,8 @@ ggplot_scatter_facet <-
            y_title = "[Y title]",
            caption = "",
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -662,6 +680,15 @@ ggplot_scatter_facet <-
     if (!is.numeric(x_var_vector)) stop("Please use a numeric x variable for a scatterplot")
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a scatterplot")
     if (is.numeric(facet_var_vector)) stop("Please use a categorical facet variable for a scatterplot")
+    
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
+    }
     
     if (is.null(pal)) pal <- pal_snz
     
@@ -901,8 +928,8 @@ ggplot_scatter_col_facet <-
            caption = "",
            legend_labels = NULL,
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -926,6 +953,15 @@ ggplot_scatter_col_facet <-
     if (!is.numeric(x_var_vector)) stop("Please use a numeric x variable for a scatterplot")
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a scatterplot")
     if (is.numeric(facet_var_vector)) stop("Please use a categorical facet variable for a scatter plot")
+    
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
+    }
     
     if (is.null(col_method)) {
       if (!is.numeric(col_var_vector)) col_method <- "category"