+++ inst/shiny/template1/global.R
@@ -7,5 +7,7 @@ library(simplevis)
 
 # load data
 ### read your RDS data in from data folder here and call it df ###
-df1 <-  ggplot2::diamonds
-df2 <- simplevis::example_sf_nz_river_wq
+df <-  ggplot2::diamonds
+
+color_vector <- sort(unique(df$color))
+