+++ inst/shiny/template1/server.R
@@ -8,9 +8,12 @@ shinyServer(function(input, output, session) {
   plot_data <- reactive({ # create a reactive data object
 
     ### add your plot_data code here ###
-    plot_data <- ggplot2::diamonds %>%
+    selected_color <- input$plot_color
+    
+    plot_data <- df %>%
+      filter(color == selected_color) %>% 
       mutate(cut = stringr::str_to_sentence(cut)) %>%
-      group_by(cut, clarity) %>%
+      group_by(cut, clarity, .drop = FALSE) %>%
       summarise(average_price = mean(price)) %>%
       mutate(average_price_thousands = round(average_price / 1000, 1)) %>%
       ungroup()
@@ -19,87 +22,88 @@ shinyServer(function(input, output, session) {
   })
 
   plot <- reactive({ # create a reactive ggplot object
-    # if (input$isMobile == F) {
-    #   font_size_title <- 11
-    #   font_size_body <- 10
-    # }
-    # else if (input$isMobile == T) {
-    #   font_size_title <- 16
-    #   font_size_body <- 15
-    # }
 
     ### add your plot code here ###
-    ### remember to add the following arguments to simplevis functions:
-          ### isMobile = input$isMobile, font_size_title = font_size_title, font_size_body = font_size_body
-    ### remember to refer to a reactive plot_data object as plot_data()
+    # remember to add isMobile = input$isMobile to simplevis functions, so that mobile plot looks good
+    # remember to refer to a reactive plot_data object as plot_data()
+    selected_color <- input$plot_color
+    
+    title <- paste0("Average diamond price of colour ", selected_color, " by cut and clarity")
+      
     plot <- ggplot_hbar_col(data = plot_data(), 
                             x_var = average_price_thousands, 
                             y_var = cut, 
                             col_var = clarity, 
                             legend_ncol = 4,
-                            title = "Average diamond price by cut and clarity", 
+                            title = title, 
                             x_title = "Average price ($US thousands)", 
                             y_title = "Cut", 
-                            # font_size_title = font_size_title,
-                            # font_size_body = font_size_body,
                             isMobile = input$isMobile)
     
 
     return(plot)
   })
 
-  output$plot_desktop <- plotly::renderPlotly({ ### render it as a html object for desktop users
+  output$plot_desktop <- plotly::renderPlotly({ 
     plotly::ggplotly(plot(), tooltip = "text") %>%
       simplevis::remove_plotly_buttons()
   })
 
-  output$plot_mobile <- renderPlot({ ### render it as a image for mobile users
-    plot() +
-      ggplot2::theme(plot.title.position = "plot") +
-      ggplot2::theme(plot.caption.position = "plot") 
-  })
-  
-  ### use this code if you want to speed things up for mobile users
-  # output$plot_mobile <- renderCachedPlot({
-  #   plot()
-  # },
-  # cacheKeyExpr = {
-  #   list() ### list input$dependencies here
+  # output$plot_mobile <- renderPlot({ 
+  #   plot() +
+  #     ggplot2::theme(plot.title.position = "plot") +
+  #     ggplot2::theme(plot.caption.position = "plot") 
   # })
   
+  ### use this code if you want to speed things up for mobile users
+  output$plot_mobile <- renderCachedPlot({
+    plot() +
+        ggplot2::theme(plot.title.position = "plot") +
+        ggplot2::theme(plot.caption.position = "plot")
+  },
+  cacheKeyExpr = {
+    list() ### list input$dependencies here
+  })
+  
   ### table ###
-
+  
+  # table_data <- reactive({   #if more than one dataset
+  #   if(input$table_data == "Diamonds") table_data <- ggplot2::diamonds
+  #   else if(input$table_data == "Storms") table_data <- dplyr::storms
+  #   return(table_data)
+  # })
+  
   output$table <- DT::renderDT(
     df, ### adjust data object name, and columns as necessary ###
+    # table_data(), 
     filter = "top",
     rownames = F,
-    options = list(pageLength = 5,
-                   scrollX = T)
+    options = list(pageLength = 5, scrollX = T)
   )
 
   ### download ###
-
+  
+  # use this code if one dataset
   output$download <- downloadHandler(
-    ### add a zip file called download.zip into the data subfolder ###
-    filename <- function() {
-      "download.zip"
+    filename = function() {
+      "data.csv"
     },
-    content <- function(file) {
-      file.copy("data/download.zip", file)
-    },
-    contentType = "application/zip"
+    content = function(file) {
+      readr::write_csv(df, file, na = "") # adjust data object name, as necessary 
+    }
   )
   
+  # use this code if multiple datasets
   # output$download <- downloadHandler(
-  #   ### applicable  if 1 dataset ###
-  #   filename = function() {
-  #     "data.csv"
+  #   filename <- function() {
+  #     "data.zip"   # add a zip file called data.zip into the data subfolder
   #   },
-  #   content = function(file) {
-  #     readr::write_csv(df, file, na ="") ### adjust data object name, as necessary ###
-  #   }
+  #   content <- function(file) {
+  #     file.copy("data/data.zip", file)
+  #   },
+  #   contentType = "application/zip"
   # )
-
+  
   ### download code ###
 
   output$download_code <- downloadHandler(