+++ inst/shiny/template1/ui.R
@@ -23,11 +23,6 @@ shinyUI(
     tags$head(includeCSS("www/style.css")),
     tags$head(includeScript("www/js/tag-manager.js")),
     mobileDetect("isMobile"),
-    # https://github.com/g3rv4/mobileDetect
-    # tags$head(
-    #   tags$script(src = "https://cdnjs.cloudflare.com/ajax/libs/iframe-resizer/3.5.16/iframeResizer.contentWindow.min.js",
-    #               type = "text/javascript")
-    # ), # https://www.cultureofinsight.com/blog/2018/03/15/2018-03-15-responsive-iframes-for-shiny-apps/
     navbarPage(
       title = HTML("<b>", app_title, "</b>"),
       windowTitle = app_title,
@@ -36,8 +31,9 @@ shinyUI(
         icon = icon("chart-area", lib = "font-awesome"),
         sidebarLayout(
           sidebarPanel(
-            width = 3
+            width = 3,
             ### add radioButtons and other widgets in here ###
+            radioButtons("plot_color", "Colour", color_vector)
           ),
           mainPanel(
             width = 9,
@@ -66,6 +62,17 @@ shinyUI(
         fluidRow(
           shinycssloaders::withSpinner(DT::DTOutput("table"), type = 7, color = "#A8A8A8")
         )
+        # sidebarLayout( #use this if multiple datasets
+        #   sidebarPanel(
+        #     width = 3,
+        #     ### add radioButtons and other widgets in here ###
+        #     radioButtons("table_data", "Dataset", c("Diamonds", "Storms"))
+        #   ),
+        #   mainPanel(
+        #     width = 9,
+        #     shinycssloaders::withSpinner(DT::DTOutput("table"), type = 7, color = "#A8A8A8")
+        #   )
+        # )
       ),
       tabPanel(
         "Download",
@@ -78,6 +85,5 @@ shinyUI(
         fluidRow(includeMarkdown("www/about.Rmd"))
       )
     )
-    # HTML('<div data-iframe-height></div>') # https://www.cultureofinsight.com/blog/2018/03/15/2018-03-15-responsive-iframes-for-shiny-apps/
   )
 )